extract TodoViewHolder.kt
